requirements.txt
# Web 框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# 异步 HTTP 客户端
httpx>=0.25.0

# 数据验证
pydantic>=2.0.0

# 环境变量管理
python-dotenv>=1.0.0

# 通用工具
requests>=2.31.0
tqdm>=4.66.0

# 文件与路径处理
pathlib>=1.0.1

# 日志
loguru>=0.7.0


# 安装 dashscope
dashscope>=1.19.0

# 可选：如果你计划集成本地向量知识库（如FAISS）
# faiss-cpu>=1.7.4
# numpy>=1.24.0
# scikit-learn>=1.3.0

# 可选：如果你使用 Beautiful Soup 解析网页内容
# beautifulsoup4>=4.12.0

# 注意：
# - 图像生成（文生图）通常通过外部 API（如阿里通义万相、OpenAI DALL·E、Stability AI）实现，
#   因此无需安装 heavy 的本地生成模型（如 Stable Diffusion），只需用 requests 调用即可。
# - 3D 地图使用前端 Three.js，后端无需图形库。